Initial refactoring to fit the px-files and the binary files behind the same datasource interface.

+++ PxApi/Controllers/DataController.cs
@@ -236,7 +236,7 @@ namespace PxApi.Controllers
                 MatrixMap requestMap = MetaFiltering.ApplyToMatrixMeta(meta, query);
 
                 long maxSize = AppSettings.Active.QueryLimits.JsonStatMaxCells;
-                int size = requestMap.GetSize();
+                long size = requestMap.GetSize();
                 if (size > maxSize)
                 {
                     logger.LogInformation("Too large request received. Size: {Size}.", size);
